+++ src/shader/ribbon_patch.vert
@@ -3,14 +3,14 @@
 layout( location = 0 ) in vec3 vao_controlPointPosition;
 layout( location = 1 ) in vec3 vao_controlPointDirection;
 layout( location = 2 ) in vec3 vao_controlPointNormal;
-layout( location = 3 ) in uint vao_controlPointSecondaryStructure;
+layout( location = 3 ) in unsigned short vao_controlPointSecondaryStructure;
 layout( location = 4 ) in vec3 vao_controlPointColor;
 
-out vec3	  vs_position;
-out vec3	  vs_direction;
-out vec3	  vs_normal;
-flat out uint vs_secondaryStructure;
-flat out vec3 vs_color;
+out vec3				vs_position;
+out vec3				vs_direction;
+out vec3				vs_normal;
+flat out unsigned short vs_secondaryStructure;
+flat out vec3			vs_color;
 
 void main()
 {
